requirements.txt
Flask
flask-session
flask-cors
python-dotenv
gunicorn
requests
sentence-transformers
SQLAlchemy
msal
faiss-cpu
PyMuPDF
docx2txt
unstructured
pydantic
pytesseract
pdf2image
Pillow
